test(repositories): Extract StubJenkinsClient from test class
